requirements.txt
aiogram
cryptocompare
python-decouple
SQLAlchemy==1.4.3
aiosqlite
asyncpg
aioredis==1.3.1
loguru